[42] Move syntactic filter to separate module.
